requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
dj-database-url==2.0.0
Django==4.2.3
django-allauth==0.54.0
django-channels==0.7.0
django-cleanup==8.0.0
django-cors-headers==4.2.0
django-debug-toolbar==4.1.0
django-environ==0.10.0
django-parler==2.3
django-parler-rest==2.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
msgpack==1.0.5
oauthlib==3.2.2
Pillow==10.0.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
pymemcache==4.0.0
python-memcached==1.59
python3-openid==3.2.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.3
whitenoise==6.5.0
